defect/DE8836-jekyll-contentful-issues
This PR changes the limit back to 1000 to reduce the number of API
calls we are making. In Contentful, the Article content model had the
tags and recommended media fields disabled in the API response. As a
result, querying with limit of 1000 works again without throwing a 400
Response too Large error.